--- app/academicSecretary/eduProg/create.tsx
@@ -1,9 +1,0 @@
-import { EducationalProgramCreate } from "../../../components/features/academicSecretary/eduProg/application/screens/create";
-
-
-export default function createScreen(){
-    return(
-      <EducationalProgramCreate/>
-        
-    )
-  }
