--- app/academicSecretary/eduProg/delete.tsx
@@ -1,8 +1,0 @@
-import { EducationalProgramDelete } from "../../../components/features/academicSecretary/eduProg/application/screens/delete";
-
-export default function deleteScreen(){
-    return(
-      <EducationalProgramDelete/>
-        
-    )
-  }
