--- app/academicSecretary/eduProg/home.tsx
@@ -1,9 +1,0 @@
-import EducationalProgramsView from "../../../components/features/academicSecretary/eduProg/application/screens/home";
-
-
-export default function EducationalProgramsScreen(){
-  return(
-    <EducationalProgramsView/>
-      
-  )
-}
