--- app/academicSecretary/eduProg/update.tsx
@@ -1,8 +1,0 @@
-import { EducationalProgramUpdate } from "../../../components/features/academicSecretary/eduProg/application/screens/update";
-
-export default function updateScreen(){
-    return(
-      <EducationalProgramUpdate/>
-        
-    )
-  }
