+++ app/academicSecretary/educationalProgram/create.tsx
@@ -1,0 +1,9 @@
+import { EducationalProgramCreate } from "../../../components/features/academicSecretary/educationalPrograms/application/screens/createEducationalProgram";
+
+
+export default function createScreen(){
+    return(
+      <EducationalProgramCreate/>
+        
+    )
+  }
