+++ app/academicSecretary/educationalProgram/delete.tsx
@@ -1,0 +1,8 @@
+import { EducationalProgramDelete } from "../../../components/features/academicSecretary/educationalPrograms/application/screens/deleteEducationalProgram";
+
+export default function deleteScreen(){
+    return(
+      <EducationalProgramDelete/>
+        
+    )
+  }
