+++ app/academicSecretary/educationalProgram/home.tsx
@@ -1,0 +1,9 @@
+import EducationalProgramsView from "../../../components/features/academicSecretary/educationalPrograms/application/screens/homeEducationalProgram";
+
+
+export default function EducationalProgramsScreen(){
+  return(
+    <EducationalProgramsView/>
+      
+  )
+}
