+++ app/academicSecretary/educationalProgram/update.tsx
@@ -1,0 +1,8 @@
+import { EducationalProgramUpdate } from "../../../components/features/academicSecretary/educationalPrograms/application/screens/updateEducationalProgram";
+
+export default function updateScreen(){
+    return(
+      <EducationalProgramUpdate/>
+        
+    )
+  }
