--- app/programDirector/approvedTemplate.tsx
@@ -1,8 +1,0 @@
-import ApprovedTemplateView from "../../components/features/programDirector/application/screens/approvedTemplateView"
-
-export default function ApprovedTemplateScreen(){
-  return(
-    <ApprovedTemplateView/>
-      
-  )
-}
