--- app/programDirector/correctedTemplate.tsx
@@ -1,8 +1,0 @@
-import CorrectedTemplateView from "../../components/features/programDirector/application/screens/correctedTemplateView"
-
-export default function CorrectedTemplateScreen(){
-  return(
-    <CorrectedTemplateView/>
-      
-  )
-}
