+++ components/features/academicSecretary/educationalPrograms/application/screens/createEducationalProgram.tsx
