+++ components/features/academicSecretary/homeView.tsx
@@ -1,6 +1,6 @@
 import { Link, router } from "expo-router";
 import { Button } from "react-native-paper";
-import { globalStyles } from "../../../auth/application/screens/loginView";
+import { globalStyles } from "../auth/application/screens/loginView";
 import { ThemedView } from "@/components/ThemedView";
 
 export default function HomeView() {
